+++ doc/changelog.rst
@@ -36,6 +36,8 @@ Simple Supplier
 Order Printouts
 ~~~~~~~~~~~~~~~
 
+- Move printouts to tab from toolbar
+
 Campaigns
 ~~~~~~~~~